Rename unit test 'blocks' to avoid name collision
