+++ unit_test.py
@@ -2,6 +2,6 @@
 
 import unit_tests
 
-unit_tests.blocks.run()
+unit_tests.test_blocks.run()
 unit_tests.legos.run()
 unit_tests.primitives.run()
