+++ unit_tests/__init__.py
@@ -1,3 +1,3 @@
-import blocks
+import test_blocks
 import legos
 import primitives
